+++ frontend/ui/src/app/interfaces/models/dogma/mythology/deity.ts
@@ -1,0 +1,2 @@
+export interface Deity {
+}