+++ frontend/ui/src/app/interfaces/models/dogma/mythology/fae.ts
@@ -1,0 +1,2 @@
+export interface Fae {
+}