+++ frontend/ui/src/app/interfaces/models/places/landmark.ts
@@ -1,0 +1,2 @@
+export interface Landmark {
+}