+++ frontend/ui/src/app/interfaces/models/places/nation-type.ts
@@ -1,0 +1,2 @@
+export interface NationType {
+}