+++ frontend/ui/src/app/interfaces/models/politics/military/squad-leader.ts
@@ -1,0 +1,2 @@
+export interface SquadLeader {
+}