+++ frontend/ui/src/app/services/dogma/mythology/deity/deity.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { DeityService } from './deity.service';
+
+describe('DeityService', () => {
+  let service: DeityService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(DeityService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});