+++ frontend/ui/src/app/services/dogma/mythology/demon/demon.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { DemonService } from './demon.service';
+
+describe('DemonService', () => {
+  let service: DemonService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(DemonService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});