+++ frontend/ui/src/app/services/dogma/mythology/fae/fae.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FaeService } from './fae.service';
+
+describe('FaeService', () => {
+  let service: FaeService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(FaeService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});