+++ frontend/ui/src/app/services/dogma/prophet/prophet.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ProphetService } from './prophet.service';
+
+describe('ProphetService', () => {
+  let service: ProphetService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ProphetService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});