+++ frontend/ui/src/app/services/society/person/person.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {Observable, throwError} from "rxjs";
-import {CustomResponse} from "../../../interfaces/custom-response/custom-response";
-import {Gender} from "../../../enums/gender/gender";
+import {CustomResponse} from "../../../interfaces/custom-response";
+import {Gender} from "../../../enums/gender";
 import {catchError, tap} from "rxjs/operators";
 import {HttpClient, HttpErrorResponse} from "@angular/common/http";