+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/politics/enumeration/RankType.java
@@ -1,7 +1,0 @@
-package com.testament.veltahleon.model.entities.politics.enumeration;
-
-public enum RankType {
-    NAVAL,
-    AERIAL,
-    TERRESTRIAL
-}