+++ veltahleon/src/main/java/com/testament/veltahleon/model/enumeration/RankType.java
@@ -1,0 +1,7 @@
+package com.testament.veltahleon.model.enumeration;
+
+public enum RankType {
+    NAVAL,
+    AERIAL,
+    TERRESTRIAL
+}